modified demo_onacid_mesoscope.py
